+++ test/endpoint.js
@@ -46,10 +46,12 @@ describe('phonetics', function() {
   require('./phonetics/caverphone.js');
   require('./phonetics/daitch-mokotoff.js');
   require('./phonetics/double-metaphone.js');
+  require('./phonetics/lein.js');
   require('./phonetics/metaphone.js');
   require('./phonetics/mra.js');
   require('./phonetics/nysiis.js');
   require('./phonetics/soundex.js');
+  require('./phonetics/statcan.js');
 
   describe('french', function() {
     require('./phonetics/french/phonetic.js');